refactor: split monorepo.rs and versioning.rs into focused submodules
